+++ QufoxServer.js
@@ -2,14 +2,12 @@ var util = require('util');
 var Sockets = require('socket.io');
 var tools = require('./tools');
 
-exports.MessagingServer = (function(){
-	function MessagingServer (listenTarget, option, adapter)	{
+exports.QufoxServer = (function(){
+	function QufoxServer (listenTarget, option, adapter)	{
 		var self = this;
 
 		var io = Sockets(listenTarget, option);		
 		if (adapter) io.adapter(adapter);
-
-		//self.io.path('/messaging');
 
 		io.on('connection', function (socket){			
 			tools.log('connected - ' + ' (socketId: ' + socket.id + ' )');
@@ -35,28 +33,11 @@ exports.MessagingServer = (function(){
 				socket.emit('leaveCallback', {id:sessionId, data:'success'});
 			});
 
-			// var rpServer = new RequestResponseSocketServer(socket)
-			// .set('join', function (sessionId){								
-			// 	socket.join(sessionId);
-			// 	tools.log('joinSession -'+ ' (sessionId: ' + sessionId + ')');
-			// 	res.send({err:null, result:{talks:[]}});
-			// })
-			// .set('addNewTalk', function (req, res){
-			// 	tools.log('addNewTalk -' + util.inspect(req.data));
-			// 	sessionEvent('onNewTalk', req.data.sessionId, req.data);
-			// 	res.send({err:null, result:req.data});
-			// });
-
-			// function sessionEvent(eventName, sessionId, data){
-			// 	var eventArg = {eventName:eventName, sessionData:{sessionId:sessionId, data:data}};				
-			// 	socket.broadcast.to(sessionId).emit('sessionEvent', eventArg);
-			// }
-
 			socket.on('disconnect', function () {
 				tools.log('disconnect - ' + ' (socketId: ' + socket.id + ' )');
 			});
 		});
 	}
 
-	return MessagingServer;
+	return QufoxServer;
 })();
